refactor(ui): extract custom hook and mock data to dedicated hook file
